fix(types): resolve Permission export ambiguity in index.ts
