+++ src/types/index.ts
@@ -1,2 +1,22 @@
+export type { UserRole } from './auth';
+export type { Permission, User, CustomClaims } from './auth';
+export * from './card';
+export * from './churning';
 export * from './firebase';
-export * from './opportunity';
+export * from './roles';
+export * from './scraper';
+export * from './server-actions';
+export * from './session';
+export * from './transformed';
+export * from './trpc';
+export * from './user';
+
+export type {
+  OfferType as OpportunityOfferType,
+  Bonus as OpportunityBonus,
+  BonusRequirements as OpportunityBonusRequirements,
+  BonusTier as OpportunityBonusTier,
+  Details as OpportunityDetails,
+} from './opportunity';
+
+export type { Permission as AuthPermission, User as AuthUser } from './auth';
